Reapply "Cleanup2 (#3501)" (#3503) (#3504)
* Reapply "Cleanup2 (#3501)" (#3503)

This reverts commit c624d6a868992619eaf5040239b53e1a94e190f4.

* Update DiscordOpaqueTokenIntrospector.java

* Update PromissoryNoteModel.java

* Revert "Update PromissoryNoteModel.java"

This reverts commit 4c5fa6ae6bd214a62876043bbeabcbad4e36c44b.

* Refactor static access and minor logic improvements

Refactored several static method calls in Mapper to use direct field access instead of getter methods for consistency and clarity. Marked BotLogger as a utility class, made DISCORD_RATE_LIMIT private, and improved error handling logic. In PromissoryNoteModel, fixed color comparison logic and corrected promissory note owner lookup to use alias.

* F

+++ src/test/java/ti4/model/StrategyCardSetModelTest.java
@@ -33,6 +33,6 @@ public class StrategyCardSetModelTest extends BaseTi4Test {
                         .map(StrategyCardModel::getInitiative)
                         .distinct()
                         .count()
-                == (long) scSetModel.getScIDs().size();
+                == scSetModel.getScIDs().size();
     }
 }
